Restructured file structure a bit Made a main.py to fix a bug causeing the program to run Took out swiping feature until its fixed
