+++ kaleidoscopic-kings/frontend/swipe.py
@@ -4,39 +4,19 @@
 # from kivy.uix.boxlayout import BoxLayout
 # from kivy.uix.image import AsyncImage
 from kivy.core.window import Window
-# from kivy.app import App
-# from kivy.uix.widget import Widget
 from kivy.animation import Animation
-from kivy.animation import AnimationTransition
-from kivy.uix.carousel import Carousel
 from kivy.uix.scatter import Scatter
+import frontend.animation_controller as ac
 
 
-KEYFRAME = 0
-
-
-class Trans(AnimationTransition):
-    @staticmethod
-    def trans(progress):
-        global KEYFRAME
-        return max(0, min(KEYFRAME, 1))
-
-    def norm(progress):
-        return progress + progress
-
-
-class Swipe(Carousel):
+class SwipeCard(Scatter):
     def __init__(self, **kwargs):
-        super(Swipe, self).__init__(**kwargs)
-        # self.direction='right'
-
-
-class Card(Scatter):
-    def __init__(self, **kwargs):
-        super(Card, self).__init__(**kwargs)
+        super(SwipeCard, self).__init__(**kwargs)
+        self.init_x = self.x
+        self.init_y = self.y
         self.keyboard = Window.request_keyboard(None, self)
         self.keyboard.bind(on_key_down=self.on_keyboard_down)
-        self.trans = Trans.trans
+        self.trans = ac.Trans.trans
 
         self.saxis = 0
         self.axis = 0
@@ -45,15 +25,13 @@ class Card(Scatter):
         self.anim = None
 
     def on_touch_down(self, touch):
-        global KEYFRAME
-        KEYFRAME = 0
+        ac.KEYFRAME = 0
         self.axis = touch.spos[0]
         self.saxis = touch.spos[0]
         self.anim = None
         self.dir = 0
 
     def on_touch_move(self, touch):
-        global KEYFRAME
 
         if self.dir == 0:
             self.anim = None
@@ -64,42 +42,40 @@ class Card(Scatter):
         self.dir = tmp_dir
         if self.anim is None:
             if self.dir == -1:
-                self.anim = Animation(x=self.x-290,
-                                      scale=self.scale + 1,
-                                      rotation=90,
+                self.anim = Animation(x=self.init_x-1500,
+                                      # scale=self.scale + 1,
+                                      # rotation=90,
                                       transition=self.trans,
                                       duration=1000)
                 self.anim.start(self)
             if self.dir == 1:
-                self.anim = Animation(x=self.x+290,
-                                      scale=self.scale + 1,
-                                      rotation=-90,
+                self.anim = Animation(x=self.init_y+1500,
+                                      # scale=self.scale + 1,
+                                      # rotation=-90,
                                       transition=self.trans,
                                       duration=1000)
                 self.anim.start(self)
 
         difference = self.saxis - touch.spos[0]
-        KEYFRAME = abs(difference)
+        ac.KEYFRAME = abs(difference)
 
         self.axis = touch.spos[0]
 
     def on_touch_up(self, touch):
-        global KEYFRAME
-        KEYFRAME = 0
+        ac.KEYFRAME = 0
         self.saxis = 0
         self.axis = 0
         self.dir = 0
 
     def on_keyboard_down(self, keyboard, keycode, text, modifiers):
         """Moves the image around."""
-        global KEYFRAME
 
         anim = None
         if keycode[1] == 'left':
             anim = Animation(x=self.x-50,
                              transition=self.trans, duration=1000)
         elif keycode[1] == 'right':
-            KEYFRAME += 0.1
+            ac.KEYFRAME += 0.1
             # anim = Animation(x=self.x+50)
         elif keycode[1] == 'up':
             anim = Animation(y=self.y+50)
